Idea/package_thuvienphapluat/requirements.txt
requests
beautifulsoup4
lxml
pandas
playwright
vncorenlp
phonlp
google-generativeai
openpyxl
python-dotenv
